requirements.txt
flask
numpy
Pillow
pydicom
torch
transformers